requirements.txt
easyocr==1.7.0
img2table==1.0.6
matplotlib==3.7.2
matplotlib-inline==0.1.6
numpy==1.23.5
opencv-python==4.6.0.66
opencv-python-headless==4.8.0.74
openpyxl==3.1.2
pandas==2.0.3
Pillow==9.3.0
python==3.9.6
toml==0.10.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2